requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
email-validator
python-dotenv==1.0.0
py-sc-client>=0.3.0
py-sc-kpm>=0.3.0
langchain-ollama
langchain-core
langchain-chroma
python-jose[cryptography]
google-auth
python-multipart
sqlalchemy>=2.0
psycopg2-binary
bcrypt
